ITC 2/BroncoBook-ITC-MISSA/Application/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
openai==1.68.2
python-dotenv==1.0.1
pydantic==2.11.7
numpy==2.3.3
pytest==9.0.2
